Convert all files to ES module
