+++ examples/merge_plain_functions.mjs
@@ -1,4 +1,4 @@
-var typed = require('../typed-function');
+import typed from '../src/typed-function.mjs';
 
 // create a couple of plain functions with a signature
 function fn1 (a) {
