+++ examples/recursion.mjs
@@ -1,13 +1,13 @@
-var typed = require('../typed-function');
+import typed from '../src/typed-function.mjs';
 
 // create a typed function that invokes itself
 var sqrt = typed({
   'number': function (value) {
     return Math.sqrt(value);
   },
-  'string': function (value) {
-    return this(parseInt(value, 10));
-  }
+  'string': typed.referToSelf(self => function (value) {
+    return self(parseInt(value, 10));
+  })
 });
 
 // use the typed function
