+++ package.json
@@ -4,6 +4,7 @@
   "description": "Type checking for JavaScript functions",
   "author": "Jos de Jong <wjosdejong@gmail.com> (https://github.com/josdejong)",
   "contributors": [
+    "Glen Whitney (https://github.com/gwhitney)",
     "Luke Gumbley (https://github.com/luke-gumbley)"
   ],
   "homepage": "https://github.com/josdejong/typed-function",
@@ -21,16 +22,13 @@
   "dependencies": {},
   "devDependencies": {
     "benchmark": "2.1.4",
-    "brace-expansion": "2.0.1",
     "nyc": "15.1.0",
     "mocha": "10.0.0",
-    "pad-right": "0.2.2",
-    "uglify-js": "3.16.3"
+    "pad-right": "0.2.2"
   },
-  "comment": "brace-expansion is installed because an old insecure version is used by one of the dev depencencies (under istanbul)",
-  "main": "typed-function.js",
+  "type": "module",
+  "main": "src/typed-function.mjs",
   "scripts": {
-    "build": "uglifyjs typed-function.js -o typed-function.min.js -c -m",
     "test": "mocha test --recursive",
     "coverage": "nyc _mocha -- test --recursive; echo \"\nCoverage report is available at ./coverage/lcov-report/index.html\"",
     "prepublishOnly": "npm test && npm run build"
