--- test/browser.min.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>browser test - minified file</title>
-  <script src="../typed-function.min.js"></script>
-</head>
-<body>
-<script>
-  // create a typed function
-  var fn1 = typed({
-    'number, string': function (a, b) {
-      return 'a is a number, b is a string';
-    }
-  });
-
-  // use the function
-  document.write(fn1(2, 'foo') + '<br>'); // outputs 'a is a number, b is a string'
-</script>
-</body>
-</html>
