+++ test/errors.test.mjs
@@ -1,5 +1,5 @@
-var assert = require('assert');
-var typed = require('../typed-function');
+import assert from 'assert';
+import typed from '../src/typed-function.mjs';
 
 describe('errors', function () {
   it('should give correct error in case of too few arguments (named function)', function() {
